+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Print start time
+- Print exercise numbers
+
 ## [1.1.0] - 2023-01-15
 
 ### Added