+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2024-01-17
+
+### Added
+
+- Multiplication and division operations
+
 ## [1.4.0] - 2024-01-13
 
 ### Added